requirements.txt
praw
forex-python
python-dotenv
